Moved window related stuff to slWindow class
